Fixed bug; Added features
Fixed program exit when window is set to close to the system tray.
Open Directory menu will highlight the file in the opened Explorer window.
Added option to set a downloaded file's date and time from the server's Last-Modified header value.
Added detection of an extensionless "portable" file (in the executable's directory) to enable portable mode.